feat : report_api connection

See merge request s12-final/S12P31S209!140

+++ Android/app/src/main/java/com/example/myapplication/data/repository/CareRepository.kt
@@ -43,7 +43,7 @@ class CareRepository @Inject constructor(
     }
 
     suspend fun getReportDetail(
-        reportId :Int
+        reportId :Long
     ):Response<ReportDetailDto>{
         return careService.getReportDetail(reportId)
     }
